Nearly Complete.

+++ BookStore/Controllers/AdminController.cs
@@ -1,5 +1,7 @@
-﻿using BookStore.Models;
+﻿using BookStore.Common;
+using BookStore.Models;
 using BookStore.Services;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using MongoDB.Driver;
 using System;
@@ -19,7 +21,7 @@ namespace BookStore.Controllers
             this.adminService = adminService;
         }
 
-        public IActionResult Login()
+        public IActionResult Login(string encryptEmail)
         {
             return View();
         }
@@ -33,7 +35,10 @@ namespace BookStore.Controllers
 
                 if(result!=null)
                 {
-                    return RedirectToAction("Index", "Home");
+                     HttpContext.Session.SetString("userName", admin.Username.ToString());
+                    string userName = HttpContext.Session.GetString("userName");
+                    var encryptEmail = Base64.Base64Encode(userName);
+                    return RedirectToAction("AllBooks","Admin");
                 }
             }
             //TODO: Validation message need to be shown
@@ -41,18 +46,25 @@ namespace BookStore.Controllers
 
         }
 
-        public IActionResult AddBook()
+        public IActionResult AddBook(string encryptEmail)
         {
-           
-
-            return View( new BookModel());
+            if(encryptEmail==null)
+            {
+                return RedirectToAction("Login");
+            }
+            return View(new BookModel());
         }
 
         [HttpPost]
-        public IActionResult AddBook(BookModel book)
+        public IActionResult AddBook(BookModel book, string encryptEmail)
         {
+            if (encryptEmail == null)
+            {
+                return RedirectToAction("Login");
+            }
             if (ModelState.IsValid)
             {
+                book.QuantityLeft = book.Quantity;
                 adminService.AddBook(book);
 
                 return RedirectToAction(nameof(AllBooks));
@@ -64,8 +76,12 @@ namespace BookStore.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public IActionResult Delete(string id)
+        public IActionResult Delete(string id, string encryptEmail)
         {
+            if (encryptEmail == null)
+            {
+                return RedirectToAction("Login");
+            }
             if (ModelState.IsValid)
             {
                 adminService.DeleteBook(id);
@@ -74,7 +90,7 @@ namespace BookStore.Controllers
 
         }
 
-        public IActionResult AllBooks()
+        public IActionResult AllBooks(string encryptEmail)
         {
             
             var books = adminService.GetAllBooks();
@@ -82,30 +98,100 @@ namespace BookStore.Controllers
             return View(books);
         }
 
-        public IActionResult SearchByISBN()
+        public IActionResult SearchByISBN(string encryptEmail)
         {
+            if (encryptEmail == null)
+            {
+                return RedirectToAction("Login");
+            }
             return View();
         }
 
         [HttpPost]
-        public IActionResult SearchByISBN(string isbn)
+        public IActionResult SearchByISBN(string isbn, string encryptEmail)
         {
+            if (encryptEmail == null)
+            {
+                return RedirectToAction("Login");
+            }
             if (!string.IsNullOrEmpty(isbn))
             {
-                try
+                if (!adminService.FindBook(isbn))
                 {
-                    var book = adminService.GetBookByIsbn(isbn);
-                    return View("BookDetails", book);
-                }
-                catch (Google.GoogleApiException ex)
-                {
-                    ModelState.AddModelError("isbn", "Error fetching book details: " + ex.Message);
+                    try
+                    {
+                        var book = adminService.GetBookByIsbn(isbn);
+                        if (book != null)
+                        {
+                            BookModel item = new BookModel
+                            {
+                                Author = book.Authors?.FirstOrDefault(),
+                                Description = book.Description,
+                                Title = book.Title,
+                                Genre = book.Categories?.FirstOrDefault(),
+                                ISBN = isbn,
+                                Image = book.ImageLinks?.Thumbnail,
+                                PageCount = book.PageCount,
+                                Publisher = book.Publisher,
+                                Quantity = 100,
+                                QuantityLeft = 100
+
+
+                            };
+
+                            return View("AddBookAPI", item);
+                        }
+                    }
+                    catch (Google.GoogleApiException ex)
+                    {
+                        ModelState.AddModelError("isbn", "Error fetching book details: " + ex.Message);
+                    }
                 }
             }
 
             return View();
         }
 
+
+        [HttpGet]
+        public IActionResult AllOrder()
+        {
+            var o = adminService.AllOrder();
+
+            List<UserOrder> userOrder = new List<UserOrder>();
+            foreach (var order in o)
+            {
+                List<string> productTitles = order.Items.Products.Select(p => p.Title).ToList();
+                var totalTitles = productTitles.Count();
+
+                foreach (var title in productTitles)
+                {
+                    UserOrder userOrderItem = new UserOrder
+                    {
+
+                        OrderDate = order.OrderDate,
+                        Title = title,
+                        TotalAmount = order.TotalAmount
+                    };
+                    userOrder.Add(userOrderItem);
+                }
+
+
+            }
+
+            return View(userOrder);
+        }
+
+        [HttpGet]
+        public IActionResult LogOut()
+        {
+
+          
+            HttpContext.Session.Remove("userName");
+            HttpContext.Session.Clear();
+
+            return RedirectToRoute(new { controller = "Home", action = "Index" });
+        }
 
     }
 }
